+++ Sources/Swift/Integrations/WatchdogTerminations/SentryWatchdogTerminationTrackingIntegration.swift
@@ -3,46 +3,59 @@ import Foundation
 
 #if (os(iOS) || os(tvOS) || targetEnvironment(macCatalyst) || os(visionOS)) && !SENTRY_NO_UIKIT
 
-typealias WatchdogTerminationTrackingProvider = ANRTrackerBuilder & ProcessInfoProvider & AppStateManagerProvider & WatchdogTerminationScopeObserverBuilder & WatchdogTerminationTrackerBuilder
+typealias WatchdogTerminationTrackingProvider = ProcessInfoProvider & AppStateManagerProvider & WatchdogTerminationScopeObserverBuilder & WatchdogTerminationTrackerBuilder & WatchdogTerminationHangTrackerBuilder
 
-final class SentryWatchdogTerminationTrackingIntegration<Dependencies: WatchdogTerminationTrackingProvider>: NSObject, SwiftIntegration, SentryANRTrackerDelegate {
+final class SentryWatchdogTerminationTrackingIntegration<Dependencies: WatchdogTerminationTrackingProvider>: NSObject, SwiftIntegration {
 
-    private let tracker: SentryWatchdogTerminationTracker
-    private let anrTracker: SentryANRTracker
     private let appStateManager: SentryAppStateManager
+    private let terminationTracker: SentryWatchdogTerminationTracker
+    private let hangTracker: SentryWatchdogTerminationHangTracker
 
     init?(with options: Options, dependencies: Dependencies) {
         guard options.enableWatchdogTerminationTracking else {
             SentrySDKLog.debug("Not going to enable \(Self.name) because enableWatchdogTerminationTracking is disabled.")
             return nil
         }
-
         guard options.enableCrashHandler else {
             SentrySDKLog.debug("Not going to enable \(Self.name) because enableCrashHandler is disabled.")
             return nil
         }
-
         guard dependencies.processInfoWrapper.environment["XCTestConfigurationFilePath"] == nil else {
             SentrySDKLog.debug("Not going to enable \(Self.name) because XCTestConfigurationFilePath is set.")
             return nil
         }
-
         guard let terminationTracker = dependencies.getWatchdogTerminationTracker(options) else {
             SentrySDKLog.fatal("Watchdog Termination tracker not available")
             return nil
         }
+        self.terminationTracker = terminationTracker
 
-        tracker = terminationTracker
-        anrTracker = dependencies.getANRTracker(options.appHangTimeoutInterval)
         appStateManager = dependencies.appStateManager
+        guard let hangTracker = dependencies.getWatchdogTerminationHangTracker(
+            timeoutInterval: options.appHangTimeoutInterval,
+            hangStarted: { [weak appStateManager] in
+                SentrySDKLog.debug("App hang detected in watchdog termination tracking")
+                appStateManager?.updateAppState { appState in
+                    appState.isANROngoing = true
+                }
+            },
+            hangStopped: { [weak appStateManager] in
+                appStateManager?.updateAppState { appState in
+                    appState.isANROngoing = false
+                }
+            }
+        ) else {
+            SentrySDKLog.fatal("Watchdog Termination tracker not available")
+            return nil
+        }
+        self.hangTracker = hangTracker
 
         super.init()
 
-        tracker.start()
-        anrTracker.add(listener: self)
+        terminationTracker.start()
+        hangTracker.start()
 
         let scopeObserver = dependencies.getWatchdogTerminationScopeObserverWithOptions(options)
-
         SentrySDKInternal.currentHub().configureScope { outerScope in
             // Add the observer to the scope so that it can be notified when the scope changes.
             outerScope.add(scopeObserver)
@@ -67,22 +80,8 @@ final class SentryWatchdogTerminationTrackingIntegration<Dependencies: WatchdogT
     }
 
     func uninstall() {
-        tracker.stop()
-        anrTracker.remove(listener: self)
-    }
-
-    // MARK: - SentryANRTrackerDelegate
-
-    func anrDetected(type: SentryANRType) {
-        appStateManager.updateAppState { appState in
-            appState.isANROngoing = true
-        }
-    }
-
-    func anrStopped(result: SentryANRStoppedResult?) {
-        appStateManager.updateAppState { appState in
-            appState.isANROngoing = false
-        }
+        terminationTracker.stop()
+        hangTracker.stop()
     }
 }